Update Average.h
I suggest two modifications to accelerate the program:
(1) The use of a circular buffer to accelerate the push function.
(2) Calculating the running sum to accelerate the mean determination.

This is a quick 10 minute hack, that allowed me to track 1-2 kHz trigger frequencies by tracking 10 pulse-pulse distances with an Arduino. The original library was a bit too slow.
I didn't check all the functionality of the library, so please use carefully.