Update Average.h
This is a quick hack to accelerate the library:
(1) Use of a circular buffer to store the values. This removes the requirement for shifting values in push().
(2) Continuous calculation of the sum to accelerate the mean calculation.

This is a quick hack that allowed me to track/lock 8 kHz trigger frequencies with an Arduino, by comparing interrupt pulse-pulse distances. The original library was a bit slow.
I didn't thoroughly check the library, so please use carefully.